+++ tests/Resources/MessageTest.php
@@ -76,10 +76,41 @@ class MessageTest extends BaseTestCase {
         self::assertNotNull($actual);
     }
 
+    public function testwhatsappTemplateMessageCreate()
+    {
+        $request = new PlivoRequest(
+            'POST',
+            'Account/MAXXXXXXXXXXXXXXXXXX/Message/',
+            [
+                "dst" => "+919012345678",
+                "text" => "Test",
+                "src" => "+919999999999"
+            ]);
+        $body = file_get_contents(__DIR__ . '/../Mocks/messageSendResponse.json');
+
+        $this->mock(new PlivoResponse($request,200, $body));
+
+        $template = '{
+            "name": "hello_world",
+            "language": "en_US",
+            "components": null
+          }';
+
+        $actual = $this->client->messages->create([ "src" => "+919999999999", "dst" => "+919012345678", "template"  =>$template]);
+
+        self::assertNotNull($actual);
+    }
+
     public function testMessageGet()
     {
         $messageUuid = "5b40a428-bfc7-4daf-9d06-726c558bf3b8";
         $requesterIP = "192.168.1.1";
+        $dltEntityID = "1234";
+        $dltTemplateID = "5678";
+        $dltTemplateCategory = "service_implicit";
+        $conversationID = "9876";
+        $conversationOrigin = "utility";
+        $conversationExpirationTimestamp = "2023-08-03 23:02:00+05:30";
         $request = new PlivoRequest(
             'GET',
             'Account/MAXXXXXXXXXXXXXXXXXX/Message/'.$messageUuid.'/',
@@ -96,6 +127,12 @@ class MessageTest extends BaseTestCase {
 
         self::assertEquals($actual->messageUuid, $messageUuid);
         self::assertEquals($actual->requesterIP, $requesterIP);
+        self::assertEquals($actual->dltEntityID, $dltEntityID);
+        self::assertEquals($actual->dltTemplateID, $dltTemplateID);
+        self::assertEquals($actual->dltTemplateCategory, $dltTemplateCategory);
+        self::assertEquals($actual->conversationID, $conversationID);
+        self::assertEquals($actual->conversationOrigin, $conversationOrigin);
+        self::assertEquals($actual->conversationExpirationTimestamp, $conversationExpirationTimestamp);
     }
 
     public function testMessageGetwithPowerpack()
@@ -143,6 +180,12 @@ class MessageTest extends BaseTestCase {
     {
         $requesterIP1 = "192.168.1.1";
         $requesterIP2 = "192.168.1.20";
+        $dltEntityID = "9876";
+        $dltTemplateID = "5432";
+        $dltTemplateCategory = "promotional";
+        $conversationID = "1234";
+        $conversationOrigin = "service";
+        $conversationExpirationTimestamp = "2023-08-03 23:02:00+05:30";
         $request = new PlivoRequest(
             'Get',
             'Account/MAXXXXXXXXXXXXXXXXXX/Message/',
@@ -158,6 +201,23 @@ class MessageTest extends BaseTestCase {
         self::assertNotNull($actual);
         self::assertEquals($actual->resources[0]->requesterIP, $requesterIP1);
         self::assertEquals($actual->resources[19]->requesterIP, $requesterIP2);
+
+        self::assertEquals($actual->resources[0]->dltEntityID, $dltEntityID);
+        self::assertEquals($actual->resources[0]->dltTemplateID, $dltTemplateID);
+        self::assertEquals($actual->resources[0]->dltTemplateCategory, $dltTemplateCategory);
+
+        self::assertEquals($actual->resources[0]->conversationID, $conversationID);
+        self::assertEquals($actual->resources[0]->conversationOrigin, $conversationOrigin);
+        self::assertEquals($actual->resources[0]->conversationExpirationTimestamp, $conversationExpirationTimestamp);
+        
+        self::assertObjectNotHasAttribute('dltEntityID', $actual->resources[19]);
+        self::assertObjectNotHasAttribute('dltTemplateID', $actual->resources[19]);
+        self::assertObjectNotHasAttribute('dltTemplateCategory', $actual->resources[19]);
+
+        self::assertObjectNotHasAttribute('conversationID', $actual->resources[19]);
+        self::assertObjectNotHasAttribute('conversationOrigin', $actual->resources[19]);
+        self::assertObjectNotHasAttribute('conversationExpirationTimestamp', $actual->resources[19]);
+        
     }
 
 }