+++ CHANGELOG.md
@@ -1,10 +1,20 @@
 # Change Log
-## [4.53.0](https://github.com/plivo/plivo-php/tree/v4.53.0) (2023-08-02)
+## [4.55.0](https://github.com/plivo/plivo-php/tree/v4.55.0) (2023-08-10)
 **Feature - Verify**
 - Added Create Session API
 - Added Get Session API
 - Added List Session API
 - Added Validate Session API
+
+## [4.54.0](https://github.com/plivo/plivo-php/tree/v7.33.0) (2023-08-07)
+**Feature - WhatsApp message support**
+- Added new param `template` and  new message_type `whatsapp` to [send message API](https://www.plivo.com/docs/sms/api/message#send-a-message)
+- Added  new  `message_states` (`read`)   `message_type`(`whatsapp`),`conversation_id`, `conversation_origin`, `conversation_expiry_timestamp` in [list all messages API](https://www.plivo.com/docs/sms/api/message#list-all-messages) and [get message details API](https://www.plivo.com/docs/sms/api/message#retrieve-a-message) response
+
+## [4.53.0](https://github.com/plivo/plivo-php/tree/v4.53.0) (2023-08-03)
+**Feature - DLT parameters**
+- Added new params `DLTEntityID`, `DLTTemplateID`, `DLTTemplateCategory` to the [send message API](https://www.plivo.com/docs/sms/api/message/send-a-message/)
+- Added new params `DLTEntityID`, `DLTTemplateID`, `DLTTemplateCategory` to the response for the [list all messages API](https://www.plivo.com/docs/sms/api/message/list-all-messages/) and the [get message details API](https://www.plivo.com/docs/sms/api/message#retrieve-a-message)
 
 ## [4.52.0](https://github.com/plivo/plivo-php/tree/v4.52.0) (2023-07-18)
 -  Removed object_id and object_type in the parameter as well as response in [list all numbers API]